bug fixing

+++ CMNetTrainer.py
@@ -240,7 +240,7 @@ class CMNetTrainer:
 
     def roic2bbox(self, coord):
         scale = tf.to_float(self.pre_processor.target_shape[:2] / self.model.sc_factor)
-        coord = coord/scale
+        coord = tf.to_float(coord/scale)
         bbox = tf.tile(coord, [1, 1, 2])
         bbox += [-0.02, -0.02, 0.02, 0.02]
         return bbox
